Add chunk type tests
I've added a number of chunk type tests. The tests make sure that the
`is_critical`, `is_private`, `reserved_set`, and `safe_to_copy`
functions are implemented correctly. They also ensure that all of the
defined chunk types are implemented correctly. I find it unlikely that
this part will ever regress but it is very quick to run.

There is also a test to document the output of the `fmt::Debug`
implementation and make sure it doesn't regress. The `pngcheck` example
had regressed because of changes to the output, and was fixed in pull
request #319.